refactor: remove old parse method
