+++ src/config_files.rs
@@ -1,5 +1,5 @@
-use crate::args_format::EscapeMode;
 use crate::defaults::default_quote;
+use crate::parser::EscapeMode;
 use crate::tasks::Task;
 use crate::types::DynErrResult;
 use crate::utils::{get_path_relative_to_base, get_task_dependency_graph, read_env_file};
