+++ src/lib.rs
@@ -1,6 +1,5 @@
 extern crate core;
 
-pub mod args_format;
 pub mod cli;
 pub mod config_files;
 mod defaults;
